+++ index.php
@@ -1,20 +1,6 @@
 <?php
-
-class Pessoa {
-	public $nome;
-	public $idade;
-	public $altura;
-
-	public function falarNome(){
-		echo $this->nome;
-	}
-	public function falarIdade(){
-		echo $this->idade;
-	}
-	public function falarAltura(){
-		echo $this->altura;
-	}
-}
+ 
+ require "Pessoa.php";
 
 $uma_pessoa = new Pessoa;
 $uma_pessoa->nome = "Elian";
